pyproject.toml
[tool.poetry]
name = "pre-intreview-task"
version = "0.1.0"
description = ""
authors = ["Ahmed Samir"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "2"
djangorestframework = "3.10"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
